+++ lifter/lifter.cpp
@@ -12,7 +12,7 @@
 
 // new datatype for BBInfo? Something like a domtree
 vector<BBInfo> added_blocks_addresses;
-uintptr_t original_address = 0;
+uint64_t original_address = 0;
 
 // consider having this function in a class, later we can use multi-threading to
 // explore different paths
@@ -27,7 +27,7 @@ void asm_to_zydis_to_lift(IRBuilder<>& builder, ZyanU8* data,
         while (blockAddresses->size() > 0) {
 
             runtime_address = get<0>(blockAddresses->back());
-            uintptr_t offset = FileHelper::address_to_mapped_address(
+            uint64_t offset = FileHelper::address_to_mapped_address(
                 (void*)file_base, runtime_address);
 
             debugging::doIfDebug([&]() {
@@ -102,7 +102,7 @@ void asm_to_zydis_to_lift(IRBuilder<>& builder, ZyanU8* data,
 }
 
 void InitFunction_and_LiftInstructions(ZyanU64 runtime_address,
-                                       uintptr_t file_base) {
+                                       uint64_t file_base) {
     ZydisDecoder decoder;
     ZydisFormatter formatter;
 
@@ -209,12 +209,12 @@ int main(int argc, char* argv[]) {
     auto dosHeader = (win::dos_header_t*)fileBase;
     auto ntHeaders = (win::nt_headers_x64_t*)(fileBase + dosHeader->e_lfanew);
     auto ADDRESS = ntHeaders->optional_header.image_base;
-    uintptr_t RVA = static_cast<uintptr_t>(startAddr - ADDRESS);
-    uintptr_t fileOffset = FileHelper::RvaToFileOffset(ntHeaders, RVA);
+    uint64_t RVA = static_cast<uint64_t>(startAddr - ADDRESS);
+    uint64_t fileOffset = FileHelper::RvaToFileOffset(ntHeaders, RVA);
     uint8_t* dataAtAddress = fileBase + fileOffset;
     cout << hex << "0x" << (int)*dataAtAddress << endl;
     original_address = ADDRESS;
-    cout << "address: " << ADDRESS << " filebase: " << (uintptr_t)fileBase
+    cout << "address: " << ADDRESS << " filebase: " << (uint64_t)fileBase
          << " fOffset: " << fileOffset << " RVA: " << RVA << endl;
 
     funcsignatures::search_signatures(fileData);
@@ -226,7 +226,7 @@ int main(int argc, char* argv[]) {
     long long ms = timer::getTimer();
     cout << "\n" << dec << ms << " milliseconds has past" << endl;
 
-    InitFunction_and_LiftInstructions(startAddr, (uintptr_t)fileBase);
+    InitFunction_and_LiftInstructions(startAddr, (uint64_t)fileBase);
     long long milliseconds = timer::stopTimer();
     cout << "\n" << dec << milliseconds << " milliseconds has past" << endl;
     cout << "Executed " << debugging::increaseInstCounter() - 1
