+++ lifter/utils.h
@@ -11,16 +11,16 @@ namespace FileHelper {
     win::section_header_t*
     GetEnclosingSectionHeader(uint32_t rva, win::nt_headers_x64_t* pNTHeader);
 
-    uintptr_t secCharacteristics(win::nt_headers_x64_t* ntHeaders,
+    uint64_t secCharacteristics(win::nt_headers_x64_t* ntHeaders,
                                  uint32_t rva);
 
-    uintptr_t getSectionCharacteristics(void* fileBase, uintptr_t rva);
+    uint64_t getSectionCharacteristics(void* fileBase, uint64_t rva);
 
-    uintptr_t RvaToFileOffset(win::nt_headers_x64_t* ntHeaders, uint32_t rva);
+    uint64_t RvaToFileOffset(win::nt_headers_x64_t* ntHeaders, uint32_t rva);
 
-    uintptr_t address_to_mapped_address(void* fileBase, uintptr_t rva);
+    uint64_t address_to_mapped_address(void* fileBase, uint64_t rva);
 
-    uintptr_t fileOffsetToRVA(uintptr_t fileAddress);
+    uint64_t fileOffsetToRVA(uint64_t fileAddress);
 
     void setFileBase(void* base);
